+++ e_metrobus/navigation/constants.py
@@ -82,3 +82,5 @@ DATA_SOURCES = [
     'Umweltbundesamt, "Vergleich der durchschnittlicher Emissionen einzelner Verkehrsmittel im Personenverkehr in Deutschland - Bezugsjahr 2018", 01/2020',
     'Umweltbundesamt, "Entwicklung der spezifischen Kohlendioxid-Emissionen des deutschen Strommix in den Jahren 1990 - 2019", 13/2020',
 ]
+
+POSTHOG_EVENTS = ("shared", "sources", "english")